add SearchStore
